+++ pentobi/AndroidUtils.cpp
@@ -412,6 +412,16 @@ void AndroidUtils::openSaveDialog(
 #endif
 }
 
+void AndroidUtils::quit()
+{
+#ifdef Q_OS_ANDROID
+    QAndroidJniObject::callStaticMethod<void>(
+                "java/lang/System", "exit", "(I)V", 0);
+#else
+    Qt::quit();
+#endif
+}
+
 void AndroidUtils::releasePersistableUriPermission(
         [[maybe_unused]]const QString& uri)
 {